requirements.txt
Django~=5.1.7
psycopg2-binary>=2.9,<3.0
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
python-dotenv~=1.1.0